Move `promiseUtils` to common utils
