+++ src/benchmark/framework/utils/asyncUtils/asyncScheduler.ts
@@ -1,6 +1,5 @@
+import { ResolveType } from "../../../../utils/promiseUtils";
 import { BenchmarkingLogger } from "../../logging/benchmarkingLogger";
-
-import { ResolveType } from "./promiseUtils";
 
 export class AsyncScheduler {
     private readonly schedulerLogsIdentifier: string;
