+++ src/benchmark/framework/utils/subprocessUtils/ipc/onParentProcessCallExecutor/utils.ts
@@ -2,7 +2,7 @@ import { ValidateFunction } from "ajv";
 
 import { failedAjvValidatorErrorsAsString } from "../../../../../../utils/ajvErrorsHandling";
 import { stringifyAnyValue } from "../../../../../../utils/printers";
-import { PromiseExecutor } from "../../../asyncUtils/promiseUtils";
+import { PromiseExecutor } from "../../../../../../utils/promiseUtils";
 import { IPCError } from "../ipcError";
 import { IPCMessage, createIPCErrorIPCMessage } from "../ipcProtocol";
